+++ doc/conf.py
@@ -61,7 +61,7 @@ author = u'dfis'
 # built documents.
 #
 # The short X.Y version.
-version = '0.0.1'
+version = '0.0.5'
 # The full version, including alpha/beta/rc tags.
 release = '1'