+++ templates/python.flask/{{cookiecutter.project_slug}}/config/gqc.py
@@ -5,5 +5,5 @@ WSGI_LOG = None
 
 # Flask-Log Settings
 LOG_LEVEL = 'debug'
-LOG_FILENAME = "/var/meerkat/error.log"
+LOG_FILENAME = "/var/{{cookiecutter.project_slug}}/error.log"
 LOG_ENABLE_CONSOLE = False